% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.mirtCAT.R
\name{extract.mirtCAT}
\alias{extract.mirtCAT}
\title{Extract elements from the internal person, test, and design objects}
\usage{
extract.mirtCAT(x, what)
}
\arguments{
\item{x}{either the \code{person}, \code{design}, or \code{test} object defined through a 
\code{customNextItem} definition}

\item{what}{a character vector extracting the desired element (see the Details section)}
}
\description{
This function extracts elements, as well as builds a few convienient elements, 
from the three internal \code{person}, \code{design}, or \code{test}
objects that are accessible through a \code{customNextItem} function 
definition (see \code{\link{mirtCAT}} for details).
}
\details{
Depending on which object is supplied, the following elements can be extracted.
}
\section{person}{


\describe{
   \item{\code{something}}{}
   
}
}

\section{test}{


\describe{
   \item{\code{something}}{}
   
}
}

\section{design}{


\describe{
   \item{\code{something}}{}
   
}
}
\examples{
\dontrun{

customNextItem <- function(person, design, test){
     # browser()
     
     sum(is.na(person$items_answered)) + 1L
}

}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\seealso{
\code{\link{mirt}}, \code{\link{mirtCAT}}, \code{\link{extract.mirt}}
}

