% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate_pattern}
\alias{generate_pattern}
\title{Generate a CAT pattern}
\usage{
generate_pattern(mirt_object, Theta, choices, item_answers = NULL)
}
\arguments{
\item{mirt_object}{}

\item{Theta}{}

\item{choices}{}

\item{item_answers}{}
}
\description{
Generate a CAT pattern given various inputs. Returns a character vector with length equal to
the test size.
}
\examples{
\dontrun{




}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\seealso{
\code{\link{mirtCAT}}
}

