% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_mirtCAT_env}
\alias{get_mirtCAT_env}
\title{Get the internal working environment state during mirtCAT session}
\usage{
get_mirtCAT_env(sessionName)
}
\arguments{
\item{sessionName}{the name of the session defined in \code{\link{mirtCAT}}}
}
\value{
a list containing the internal environmental components for mirtCAT
}
\description{
This function is used to access the internal state of the mirtCAT GUI session. 
It is only useful when designing a customized GUI using the \code{shinyGUI$ui}
input to \code{\link{mirtCAT}}.
}
