% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{mirtCAT-package}
\alias{mirtCAT-package}
\title{Computerized Adaptive Testing with Multidimensional Item Response Theory}
\description{
Computerized Adaptive Testing with Multidimensional Item Response Theory
}
\details{
Provides an HTML interface implemented with the shiny package to generate
adaptive and non-adaptive educational and psychological surveys and tests.
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{package}

