% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirtCAT_preamble.R
\name{mirtCAT_preamble}
\alias{mirtCAT_preamble}
\title{Preamble function called by mirtCAT}
\usage{
mirtCAT_preamble(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{mirtCAT}}}
}
\description{
This is largely an internal function called by \code{\link{mirtCAT}}, however it is made
public for better use with external web-hosting interfaces (like \url{http://www.shinyapps.io/}).
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\seealso{
\code{\link{mirtCAT}}
}

