% Generated by roxygen2 (4.0.1): do not edit by hand
\name{partial_result}
\alias{partial_result}
\title{Return the cached results from an interupted GUI session}
\usage{
partial_result()
}
\description{
If the mirtCAT GUI session was interupted, thee function will return the cached
results if they are available. If no cached results can be found, a message is printed and
a \code{NULL} object is returned.
}
\details{
Returns a list object containing the raw responses, items answered, demographic information,
and item response time.
}
\examples{
\dontrun{

partial_result()

# interupt a mirtCAT session, and run again to obtain partial information
# mirtCAT(...)
partial_result()

}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\seealso{
\code{\link{mirtCAT}}
}

