% Generated by roxygen2 (4.0.1): do not edit by hand
\name{personPlot}
\alias{personPlot}
\title{Generate plots following test completion}
\usage{
personPlot(person, pick_theta = NULL, main = "CAT Standard Errors", ...)
}
\arguments{
\item{person}{an object of class 'Person' returned from \code{\link{mirtCAT}}}

\item{pick_theta}{a number indicating which theta to plot (only applicable for multidimensional
  tests). The default is to facet each theta on one plot, but to plot only the first factor pass
  \code{pick_theta = 1}}

\item{main}{title of the plot}

\item{...}{additional arguments to be passed to \code{lattice}}
}
\description{
Generate plots following the convergence of \code{\link{mirtCAT}}.
}
\examples{
\dontrun{

person <- mirtCAT(mod, shiny_questions)
personPlot(person)

}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\seealso{
\code{\link{mirtCAT}}
}

