% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findNextItem}
\alias{findNextItem}
\title{Find next CAT item}
\usage{
findNextItem(x)
}
\arguments{
\item{x}{an object of class 'mirtCAT_design' returned from the \code{\link{mirtCAT}} function
  when passing \code{design_elements = TRUE}}
}
\value{
returns a numeric value indicating the index of the next item to be selected
}
\description{
A function that returns the next item in the computerized adaptive test.
}
\examples{
\dontrun{
# test defined in mirtCAT help file, first example
CATdesign <- mirtCAT(shiny_questions, mod, item_answers=answers, criteria = 'MI',
   design_elements = TRUE)

# returns number 1 in this case, since that's the starting item
findNextItem(CATdesign)

# detemine next item if item 1 and item 10 were answered correctly, and Theta = 0.5
CATdesign$person$responses[c(1,10)] <- 1L
CATdesign$person$items_answered[c(1,2)] <- c(1L, 10L)
CATdesign$person$thetas <- matrix(0.5)
findNextItem(CATdesign)
}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\seealso{
\code{\link{mirtCAT}}
}

