% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate_pattern}
\alias{generate_pattern}
\title{Generate a CAT pattern}
\usage{
generate_pattern(mirt_object, Theta, choices, item_answers = NULL)
}
\arguments{
\item{mirt_object}{single group object defined by the \code{mirt} package}

\item{Theta}{a numeric vector indicating the latent theta values for a single person}

\item{choices}{a list of character vectors signifying the possible choices for each item}

\item{item_answers}{(optional) a character vector indicating which of the options in
  \code{choices} should be considered the 'correct' answer. This is required for itemtypes that
  dichotomously score items (e.g., multiple choice items scored with the 3PL model)}
}
\description{
Generate a CAT pattern given various inputs. Returns a character vector with length equal to
the test size.
}
\examples{
\dontrun{

pat <- generate_pattern(mod, Theta = 0, choices = choices, item_answers=answers)

}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\seealso{
\code{\link{mirtCAT}}
}

