% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generate_pattern.R
\name{generate_pattern}
\alias{generate_pattern}
\title{Generate a CAT patterns}
\usage{
generate_pattern(mo, Theta, df = NULL)
}
\arguments{
\item{mo}{single group object defined by the \code{mirt} package}

\item{Theta}{a numeric vector indicating the latent theta values for a single person}

\item{df}{(optional) data.frame object containing questions, options, and scoring
  keys. See \code{\link{mirtCAT}} for details}
}
\description{
Generate a CAT pattern given various inputs. Returns a character vector or numeric matrix
(depending on whether a \code{df} input was supplied) with columns equal to the test size and
rows equal to the number of rows in \code{Theta}. For simulation studies, supplying a
\code{Theta} input with more than 1 row will generate a matrix of responses for
running independent CAT session when passed to \code{mirtCAT(..., local_pattern)}.
}
\examples{
\dontrun{

# return real response vector given choices and (optional) answers
pat <- generate_pattern(mod, Theta = 0, df=df)
# mirtCAT(df, mo=mod, local_pattern = pat)

# generate single pattern observed in dataset used to define mod
pat2 <- generate_pattern(mod, Theta = 0)
# mirtCAT(mo=mod, local_pattern = pat2)

# generate multiple patterns to be analyzed independently
pat3 <- generate_pattern(mod, Theta = matrix(c(0, 2, -2), 3))
# mirtCAT(mo=mod, local_pattern = pat3)

}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\seealso{
\code{\link{mirtCAT}}
}

