% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mirtCAT}
\alias{mirtCAT}
\title{Generate an adaptive or non-adaptive test HTML interface}
\usage{
mirtCAT(mirt_object, questions, ...)
}
\arguments{
\item{mirt_object}{single group object defined by the \code{mirt} package}

\item{questions}{a named list containing lists of \code{shiny} input types for each item.
Each element of the input should be a list of the form
\code{list(item = shinyInput(), answer = 'value')}. If no correct \code{answer} criteria exists
(such as in rating and Likert-scales) then this input may either be set to NA or ommited.}

\item{...}{additional arguments to pass}
}
\description{
Description
}

