% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mirtCAT.R, R/zzz-methods.R
\name{mirtCAT}
\alias{mirtCAT}
\alias{plot.mirtCAT}
\alias{print.mirtCAT}
\alias{summary.mirtCAT}
\title{Generate an adaptive or non-adaptive test HTML interface}
\usage{
mirtCAT(df = NULL, mo = NULL, method = "MAP", criteria = "seq",
  start_item = 1, local_pattern = NULL, design_elements = FALSE,
  cl = NULL, design = list(), shinyGUI = list(), preCAT = list(), ...)

\method{print}{mirtCAT}(x, ...)

\method{summary}{mirtCAT}(object, sort = TRUE, ...)

\method{plot}{mirtCAT}(x, pick_theta = NULL, SE = 1, main = NULL,
  par.strip.text = list(cex = 0.7), par.settings = list(strip.background =
  list(col = "#9ECAE1"), strip.border = list(col = "black")), ...)
}
\arguments{
\item{df}{a \code{data.frame} or \code{list} object
  containing the \code{character} vector inputs required to generate
  GUI questions through shiny. If \code{factor}s are supplied instead of \code{character} vectors
  then the inputs will be coerced using the \code{as.character()} function (set
  \code{stringsAsFactors = FALSE} when defining a \code{data.frame} to avoid this).
  Each row in the object corresponds to a unique
  item. The object supports the follow column name combinations as inputs to specify the
  type of response format, questions, options, answers, and stems:

  \describe{

    \item{\code{Type}}{Indicates the type of response input
      to use from the shiny package. The supported types are: 'radio' for radio buttons,
      'radio_inline' for radio buttons that are organized horizontally,
      'select' for a pull-down box for selecting inputs, or 'text' for requiring
      typed user input.}

    \item{\code{Question}}{If \code{df} is a \code{data.frame}, a
      character vector containing all the questions or stems to be generated.
      If \code{df} is a \code{list}, then the commands must be suitable for output with
      \code{shiny} (e.g., \code{Question <- list(h6('Item 1'), list(h6('Nested', h4(' item 2'))))})}

    \item{\code{Option.#}}{Names pertaining to the possible response
      options for each item, where the # corresponds to the specific category. For
      instance, a test with 4 unique response options for each item would contain
      the columns (\code{Option.1}, \code{Option.2}, \code{Option.3}, \code{Option.4}).
      If, however, some items have fewer categories than others then \code{NA}'s can be used for response
      options that do not apply.}

    \item{\code{Answer} or \code{Answer.#}}{(Optional) A character vector (or multiple character
      vectors) indicating the scoring key for items that have correct answer(s). If there
      is no correct answer for a question then a value of \code{NA} must be declared.}

    \item{\code{Stem}}{(Optional) a character vector of absolute or relative paths
      pointing external markdown (.md) or HTML (.html) files to be used as item stems.
      \code{NA}s are used if the item has no corresponding file.}

  }}

\item{mo}{single group object defined by the \code{mirt::mirt()} function. This is required
  if the test is to be scored adaptively or non-adaptively, but not required for general
  questionnaires. The object can be constructed by using the
  \code{\link{generate.mirt_object}} function if population parameters are known or by
  including a calibrated model estimated from the \code{\link{mirt}} function with real data.}

\item{method}{argument passed to \code{mirt::fscores()} for computing new scores in the CAT
  stage, with the addition of a \code{'fixed'} input to keep the latent trait estimates
  fixed at the previous values. Default is 'MAP'}

\item{criteria}{adaptive criteria used, default is to administer each item sequentially
  using \code{criteria = 'seq'}.

  Possible inputs for unidimensional adaptive tests include: \code{'MI'} for the maximum
  information, \code{'MEPV'} for minimum expected posterior variance,
  \code{'MLWI'} for maximum likelihood weighted information,
  \code{'MPWI'} for maximum posterior weighted information, \code{'MEI'} for
  maximum expected information, and \code{'IKLP'} as well as \code{'IKL'} for the
  integration based Kullback-Leibler criteria with and without the prior density weight,
  respectively, and their root-nitems administered weighted counter-parts, \code{'IKLn'} and
  \code{'IKLPn'}.

  Possible inputs for multidimensional adaptive tests include: \code{'Drule'}
  for the maximum determinant of the information matrix, \code{'Trule'} for the
  maximum (potentially weighted) trace of the information matrix,
  \code{'Arule'} for the minimum (potentially weighted) trace of the asymptotic covariance matrix,
  \code{'Erule'} for the  minimum value of the information matrix, and \code{'Wrule'} for
  the weighted information criteria. For each of these rules, the posterior weight for
  the latent trait scores can also be included with the \code{'DPrule'}, \code{'TPrule'},
  \code{'APrule'}, \code{'EPrule'}, and \code{'WPrule'}, respectively.
  As a safety precaution, if the
  selected criteria do not weight by the posterior (and therefore do not exist for
  extreme response styles) and less than 5 items have been administered then
  the method is temporarily switched to the posterior weighting
  until a variable response pattern is observed.

  Applicable to both unidimensional and multidimensional tests are the
  \code{'KL'} and \code{'KLn'} for point-wise Kullback-Leibler divergence and
  point-wise Kullback-Leibler with a decreasing delta value (\code{delta*sqrt(n)},
  where \code{n} is the number of items previous answered), respectively.
  The \code{delta} criteria is defined in the \code{design} object

  Non-adaptive methods applicable even when no \code{mo} is passed
  are: \code{'random'} to randomly select items, and \code{'seq'} for selecting
  items sequentially.}

\item{start_item}{two possible inputs to determine the starting item are available.
  Passing a single number will indicate the specific item to be used as the start item;
  default is 1, which selects the first item in the defined test/survey.
  If a character string is passed then the item will be selected from one of
  the item selections criteria available (see the \code{criteria} argument)}

\item{local_pattern}{a character/numeric matrix of response patterns
  used to run the CAT application without generating the GUI interface.
  This option requires complete response pattern(s) to be supplied. \code{local_pattern}
  is required to be numeric if no \code{questions} are supplied, and the responses must be
  within a valid range of the defined \code{mo} object.
  Otherwise, it must contain character values of plausible responses which corresponds to the
  answer key and/or options supplied in \code{df}}

\item{design_elements}{logical; return an object containing the test, person, and design
  elements? Primarily this is to be used with the \code{\link{findNextItem}} function}

\item{cl}{an object definition to be passed to the parallel package
  (see \code{?parallel::parLapply} for details). If defined, and if
  \code{nrow(local_pattern) > 1}, then each row will be run in parallel to help
  decrease estimation times in simulation work}

\item{design}{a list of design based control parameters for adaptive and non-adaptive tests.
  These can be

\describe{
  \item{\code{min_SEM}}{Default is \code{rep(0.3, nfact)}; minimum standard error or measurement
    to be reached for the latent traits (thetas) before the test is stopped. If the test is
    multidimensional, either a single value or a vector of values may be supplied to provide
    SEM criteria values for each dimension}

  \item{delta_thetas}{Default is \code{rep(0, nfact)}; stopping criteria based on the change in latent
    trait values (e.g., a change from \code{theta = 1.5} to \code{theta = 1.54} would
    stop the CAT if \code{delta_thetas = 0.05}). The default disables this stopping criteria}

  \item{\code{thetas.start}}{a numeric vector of starting values for the theta parameters.
    Default is \code{rep(0, nfact)}}

  \item{\code{min_items}}{minimum number of items that must be answered
    before the test is stopped. Default is \code{1}}

  \item{\code{max_items}}{maximum number of items that
    can be answered. Default is the length of the item bank}

  \item{\code{quadpts}}{Number of quadrature points used per dimension
    for integration (if required). Default is identical to scheme in \code{\link{fscores}}}

  \item{\code{theta_range}}{upper and lower range for the theta
    integration grid. Used in conjunction with \code{quadpts} to generate an equally spaced
    quadrature grid. Default is \code{c(-6,6)}}

  \item{\code{weights}}{weights used when \code{criteria == 'Wrule'}, but also
    will be applied for weighted trace functions in the T- and A-rules. The default
    weights the latent dimensions equally. Default is \code{rep(1/nfact), nfact)},
    where \code{nfact} is the number of test dimensions}

  \item{\code{KL_delta}}{interval range used when \code{criteria = 'KL'}
    or \code{criteria = 'KLn'}. Default is \code{0.1}}

  \item{\code{content}}{an optional character vector indicating the type of content measured
    by an item. Must be supplied in conjunction with \code{content_prop}}

  \item{\code{content_prop}}{an optional named numeric vector indicating the
    distribution of item content proportions. A \code{content} vector must also be supplied
    to indicate the item content membership. For instance, if \code{content} contains three
    possible item content domains 'Addition', 'Subtraction', and 'Multiplication', and the
    test should contain approximately half multiplication and a quarter of both
    addition and subtraction, then a suitable input would be

    \code{content_prop = c('Addition'=0.25, 'Subtraction'=0.25, 'Multiplication'=.5)}

    Note that \code{content_prop} must sum to 1 in order to represent valid population
    proportions.
    }

  \item{\code{classify}}{a numeric vector indicating cut-off values for classification
    above or below some prior threshold. Default does not use the classification scheme}

  \item{\code{classify_CI}}{a numeric vector indicating the confident intervals used to
    classify individuals being above or below values in \code{classify}. Values must
    be between 0 and 1 (e.g., 0.95 gives 95\% confidence interval)}

  \item{\code{exposure}}{a numeric vector specifying the amount of exposure control to apply for
    each successive item (length must equal the number of items).
    The default uses no exposure control. If the item exposure
    is greater than 1 then the \code{n} most optimal
    criteria will be randomly sampled from. For instance, if
    \code{exposure[5] == 3}, and \code{critiera = 'MI'}, then when the fifth item is to be
    selected from the remaining pool of items the top 3 candidate items demonstrating
    the largest information criteria will be sampled from. Naturally, the first and last
    elements of \code{exposure} are ignored since exposure control will be meaningless.

    If all elements in \code{exposure} are between 0 and 1 then the Sympson-Hetter exposure
    control method will be implemented. In this method, an item is administered only if it
    passes a probability simulation experiment, otherwise it is removed from the item pool.
    Values closer to 1 are more likely to appear in the test, while value closer to 0 are more
    likely to be randomly discarded.}

  \item{\code{constraints}}{A named list declaring various item selection contraints for which
    particular item, where each list element is a vector of item numbers. These include:

    \describe{
         \item{\code{not_scored}}{declaring items that can be selected but will not be used in the
           scoring of the CAT. This is primarily useful when including experimental items for
           future CATs.}
         \item{\code{independent}}{declaring which items should never appear in the same CAT session.
           Use this if, for example, item 1 and item 10 have very similar questions
           types and therefore should not appear within the same session}
         \item{\code{ordered}}{if one item is selected during the CAT, adminster this
           particular group of items in order according to the specified sequence}
         \item{\code{unordered}}{same as ordered, except the items in the group will be selected at
           random until the group is complete}
    }

  }

}}

\item{shinyGUI}{a list of GUI based parameters to be over-written. These can be

\describe{
  \item{\code{title}}{A character string for the test title. Default is
    \code{'mirtCAT'}}

  \item{\code{authors}}{A character string for the author names. Default is
    \code{'Author of survey'}. If the input is an empty string (\code{''}) then the author
    information will be omitted in the GUI}

  \item{\code{instructions}}{A three part character vector indicating how to use the GUI.
    Default is:

    \preformatted{c("Instructions:",
       "To progress through the interface, click on the action button below.",
       "Next")}
  }

  \item{\code{firstpage}}{The first page of the shiny GUI. Default prints the title
    and information message.

    \preformatted{
         list(h1('Welcome to the mirtCAT interface'),
              The following interface was created using the mirtCAT package.
              To cite the package use citation(\\'mirtCATd\\') in R.')
         }
      }

   If an empty list is passed, this page will be skipped.

  \item{\code{demographics}}{A person information page used in the GUI for collecting
    demographic information, generated using tools from the shiny package. For example,
    the following code asks the participants about their Gender:

    \preformatted{
         list(selectInput(inputId = 'gender',
                  label = 'Please select your gender.',
                  choices = c('', 'Male', 'Female', 'Other'),
                  selected = ''))
        }

     By default, the demographics page is not included.

     }

  \item{\code{demographics_inputIDs}}{a character vector required if a custom demographics
    input is used. Default is \code{demographics_inputIDs = 'gender'}, corresponding to
    the \code{demographics} default}

  \item{\code{max_time}}{maximum time allowed for the generated GUI, measured
    in seconds. For instance, if the test should stop after 10 minutes then the number
    600 should be passed (10 * 60). Default is \code{Inf}, therefore no time limit}

  \item{\code{temp_file}}{a character vector indicating where a temporary .rds file
    containing the response information should be saved while the GUI is running.
    The object will be saved after each item is successfully completed. This is used to
    save response information to the hard drive in case there are power outages or
    unexpected computer restarts.

    If \code{NULL}, no temp file will be created. Upon completion of the test, the
    temp file will be deleted}

  \item{\code{resume_file}}{a character vector indicating where a temporary .rds file
    containing the response information was saved (see \code{temp_file}). Allows the GUI
    session to be continued using the previously stored demographic and response pattern
    information. Note that the demographics GUI page will appear again, but this information
    will not be used and can be skipped.}

  \item{\code{lastpage}}{A function printing the last message, indicating that the test has been completed
    (i.e., criteria has been met). The function requires exactly one argument (called \code{person}), where
    the input argument is the person object that has been updated throught the test. The default function is

    \preformatted{function(person){
                    return(list(h5("You have successfully completed the interface.
                                   Click the action button to terminate the application.")))
                     } }
   }

   \item{\code{css}}{a character string defining CSS elements to modify the GUI presentation
     elements. The input string is passed to the argument \code{tags$style(HTML(shinyGUI$css))}
     prior to constructing the user interface}

   \item{\code{forced_choice}}{logical; require a response to each item? Default is \code{TRUE}.
     This should only be set to \code{FALSE} for surveys (not CATs)}

}}

\item{preCAT}{a list object which can be used to specify a pre-CAT block in which
  different test properties may be applied prior to beginning the CAT session. If the
  list is empty, no preCAT block will be used. All of the following elements are required
  to use the \code{preCAT} input:

  \describe{
    \item{\code{min_items}}{minimum number of items to administer before the CAT session begins.
      Default is 0}

    \item{\code{max_items}}{max number of items to administer before the CAT session begins.
      An input greater than 0 is required to run the preCAT stage}

    \item{\code{criteria}}{selection criteria (see above). Default is 'random'}

    \item{\code{method}}{estimation criteria (see above). It is generally recommended to
      select a method which can deal with all-or-none response patterns, such as 'EAP'
      or 'MAP', or in the multidimensional case 'DPrule' or 'TPrule'. Default is 'MAP'}

    \item{\code{response_variance}}{logical; terminate the preCAT stage when there is variability in the
      response pattern (i.e., when maximum-likelihood estimation contains a potential optimum)?
      Default is FALSE}
}}

\item{...}{additional arguments to be passed to \code{lattice} or \code{fscores}}

\item{x}{object of class \code{'mirtCAT'}}

\item{object}{object of class \code{'mirtCAT'}}

\item{sort}{logical; sort the response patterns based on the order they
were administered? If FALSE, the raw response patterns containing NAs will be returned
for items that were not administered}

\item{pick_theta}{a number indicating which theta to plot (only applicable for multidimensional
tests). The default is to facet each theta on one plot, but to plot only the first factor pass
\code{pick_theta = 1}}

\item{SE}{size of the standard errors to plot. The default is 1, and therefore plots the
standard error. To obtain the 95\% interval use \code{SE = 1.96} (from the z-distribution)}

\item{main}{title of the plot. Will default to \code{'CAT Standard Errors'} or
\code{'CAT ##\% Confidence Intervals'} depending on the SE input}

\item{par.strip.text}{plotting argument passed to \code{\link{lattice}}}

\item{par.settings}{plotting argument passed to \code{\link{lattice}}}
}
\value{
Returns a list object of class \code{'Person'} containing the following elements:

\describe{
  \item{\code{raw_responses}}{A numeric vector indicating the raws responses to the respective
    items, where NA indicates the item was not answered}

  \item{\code{scored_responses}}{A numeric vector of scored responses if the \code{item_answers} input
    was used for each respective item}

  \item{\code{items_answered}}{An integer vector indicating the order in which the items were
    answered}

  \item{\code{thetas}}{A numeric vector indicating the final theta estimates}

  \item{\code{SE_thetas}}{A numeric vector indicating the standard errors of the
    final theta estimates}

  \item{\code{thetas_history}}{A matrix indicating the progression of updating the theta values
    during the test}

  \item{\code{thetas_SE_history}}{A matrix indicating the standard errors for theta after each
    successive item was answered}

  \item{\code{item_time}}{A numeric vector indicating how long the respondent took to answer
    each question (in seconds)}

  \item{\code{demographics}}{A data.frame object containing the information collected on the
    first page of the shiny GUI. This is used to store the demographic information for each
    participant}

  \item{\code{classification}}{A character vector indicating whether the traits could be
    classified as 'above' or 'below' the desired cutoffs}

}
}
\description{
Provides tools to generate an HTML interface for creating adaptive and
non-adaptive educational and psychological tests using the \code{shiny} package.
Suitable for applying unidimensional and multidimensional computerized adaptive tests
using item response theory methodology. Test scoring is performed using the \code{mirt} package.
However, if no scoring is required (i.e., a standard survey) then defining a \code{mirt}
object may be omitted.
}
\details{
All tests will stop once the \code{'min_SEM'} criteria has been reached or classification
above or below the specified cutoffs can be made. If all questions should
be answered, users should specify an extremely small \code{'min_SEM'} or, equivalently,
a large \code{'min_items'} criteria to the \code{design} list input.
}
\section{HTML help files, exercises, and examples}{


To access examples, vignettes, and exercise files that have been generated with knitr please
visit \url{https://github.com/philchalmers/mirtCAT/wiki}.
}
\examples{
\dontrun{

### unidimensional scored example with generated items

# create mo from estimated parameters
set.seed(1234)
nitems <- 50
itemnames <- paste0('Item.', 1:nitems)
a <- matrix(rlnorm(nitems, .2, .3))
d <- matrix(rnorm(nitems))
dat <- simdata(a, d, 1000, itemtype = 'dich')
mod <- mirt(dat, 1)
coef(mod, simplify=TRUE)

# alternatively, define mo from population values (not run)
pars <- data.frame(a1=a, d=d)
mod2 <- generate.mirt_object(pars, itemtype='2PL')
coef(mod2, simplify=TRUE)

# simple math items
questions <- answers <- character(nitems)
choices <- matrix(NA, nitems, 5)
spacing <- floor(d - min(d)) + 1 #easier items have more variation in the options

for(i in 1:nitems){
    n1 <- sample(1:50, 1)
    n2 <- sample(51:100, 1)
    ans <- n1 + n2
    questions[i] <- paste0(n1, ' + ', n2, ' = ?')
    answers[i] <- as.character(ans)
    ch <- ans + sample(c(-5:-1, 1:5) * spacing[i,], 5)
    ch[sample(1:5, 1)] <- ans
    choices[i, ] <- as.character(ch)
}

df <- data.frame(Question=questions, Option=choices,
                              Type = 'radio', stringsAsFactors = FALSE)
head(df)

(res <- mirtCAT(df)) #collect response only (no scoring or estimating thetas)
summary(res)

# include scoring by providing Answer key
df$Answer <- answers
(res_seq <- mirtCAT(df, mod)) #sequential scoring
(res_random <- mirtCAT(df, mod, criteria = 'random')) #random
(res_MI <- mirtCAT(df, mod, criteria = 'MI', start_item = 'MI')) #adaptive, MI starting item

summary(res_seq)
summary(res_random)
summary(res_MI)

#-----------------------------------------

# run locally, random response pattern given Theta
set.seed(1)
pat <- generate_pattern(mod, Theta = 0, df=df)
head(pat)

# seq scoring with character pattern for the entire test (adjust min_items)
res <- mirtCAT(df, mod, local_pattern=pat, design = list(min_items = 50))
summary(res)

# same as above, but using special input vector that doesn't require df input
set.seed(1)
pat2 <- generate_pattern(mod, Theta = 0)
head(pat2)
print(mirtCAT(mo=mod, local_pattern=pat2))

# run CAT, and save results to object called person (start at 10th item)
person <- mirtCAT(df, mod, item_answers = answers, criteria = 'MI',
                  start_item = 10, local_pattern = pat)
print(person)
summary(person)

# plot the session
plot(person) #standard errors
plot(person, SE=1.96) #95 percent confidence intervals

#-----------------------------------------

### save response object to temp directory in case session ends early
wdf <- paste0(getwd(), '/temp_file.rds')
res <- mirtCAT(df, mod, shinyGUI = list(temp_file = wdf))

# resume test this way if test was stopped early (and temp files were saved)
res <- mirtCAT(df, mod, shinyGUI = list(resume_file = wdf))
print(res)

}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\seealso{
\code{\link{generate_pattern}}, \code{\link{generate.mirt_object}}
}
\keyword{CAT,}
\keyword{MCAT,}
\keyword{adaptive}
\keyword{computerized}
\keyword{testing}

